Implemented Jump and added/changed related tests (#68)
* Add support for Jump

According to the docs, a jump is an option with no text, only an identifier of the node to jump to.

* Link and Jump related tests now passing

The previous checkin did not address fixing the tests. The tests are now passing. Note that changes were required in order to replace "named link" with "link" and to replace unnamed link with "jump".